refactored manuscript reducers into a single file
