+++ server/routes/user.router.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const {
-    rejectUnauthenticated,
+  rejectUnauthenticated,
 } = require("../modules/authentication-middleware");
 const encryptLib = require("../modules/encryption");
 const pool = require("../modules/pool");
@@ -10,25 +10,32 @@ const router = express.Router();
 
 // Handles Ajax request for user information if user is authenticated
 router.get("/", rejectUnauthenticated, (req, res) => {
-    // Send back user object from the session (previously queried from the database)
-    res.send(req.user);
+  // Send back user object from the session (previously queried from the database)
+  res.send(req.user);
 });
 
 // Handles POST request with new user data
 // The only thing different from this and every other post we've seen
 // is that the password gets encrypted before being inserted
 router.post("/register", (req, res, next) => {
-    const username = req.body.username;
-    const password = encryptLib.encryptPassword(req.body.password);
+  const username = req.body.username;
+  const password = encryptLib.encryptPassword(req.body.password);
 
-    const queryText = `INSERT INTO "user" (username, password)
+  // First Creates the user in the database
+  const queryText = `INSERT INTO "user" (username, password)
     VALUES ($1, $2) RETURNING id`;
-    pool.query(queryText, [username, password])
-        .then(() => res.sendStatus(201))
-        .catch((err) => {
-            console.log("User registration failed: ", err);
-            res.sendStatus(500);
-        });
+
+  //Returns UserId to create shelf tied to user.
+
+  pool
+    .query(queryText, [username, password])
+    .then(() => res.sendStatus(201))
+    .catch((err) => {
+      console.log("User registration failed: ", err);
+      res.sendStatus(500);
+    });
+
+
 });
 
 // Handles login form authenticate/login POST
@@ -36,14 +43,14 @@ router.post("/register", (req, res, next) => {
 // this middleware will run our POST if successful
 // this middleware will send a 404 if not successful
 router.post("/login", userStrategy.authenticate("local"), (req, res) => {
-    res.sendStatus(200);
+  res.sendStatus(200);
 });
 
 // clear all server session information about this user
 router.post("/logout", (req, res) => {
-    // Use passport's built-in method to log out the user
-    req.logout();
-    res.sendStatus(200);
+  // Use passport's built-in method to log out the user
+  req.logout();
+  res.sendStatus(200);
 });
 
 module.exports = router;
